Refactor and reorganize user and admin services, controllers, and DTOs; remove unused classes and methods.
