+++ api/src/main/java/com/icfes_group/dto/PersonaDTO.java
@@ -33,7 +33,4 @@ public class PersonaDTO {
     
     @NotBlank(message = "El email es obligatorio")
     private String email;
-    
-    @NotNull(message = "El rol es obligatorio")
-    private Integer rol_id;
 }
