+++ api/src/main/java/com/icfes_group/model/User.java
@@ -3,6 +3,7 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
 package com.icfes_group.model;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import com.icfes_group.dto.PersonaDTO;
 import com.icfes_group.dto.UserDTO;
 import jakarta.persistence.Column;
@@ -29,6 +30,7 @@ public class User {
     private Persona persona;
 
     @Column(nullable = false)
+    @JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
     private String password;
 
     @Column(nullable = false)
