+++ api/src/main/java/com/icfes_group/service/PersonService.java
@@ -1,0 +1,8 @@
+package com.icfes_group.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class PersonService {
+        //acciones como cambiar mi información y ese tipo de cosas se harán aqui
+}
